FastSimulation.
Implemented feature allows user to move transported particles
in TGeant3 or TGeant4 to any position.
It is used when implementing fast simulation in some detector
to move particle forward out of this detector to reduce computing time.

Introduced methods of the FairGenericStack should be implemened
by individual experiments, as shown in FairStack:
FastSimMoveParticleTo(): moves the particle to a position and momentum specified by the user.
FastSimStopParticle(): kill the current particle.
FastSimGetMovedIndex(): to get the stack index of the moved particle.
FastSimClearMovedIndex(): to clear the index (set it to -2).

base/sim/fastsim develops a fast simulation model for Geant4 that uses
TG4VUserFastSimulation to implement the functionality.