Fix running Geant4 Multithreaded
Protected creating of the output tree with lock_guard
FairRunSim::InitOnWorker().
Added FairRunSim::WasMT() method to check if the simulation
was performed in MT mode.
Added comments to FairRunSim MT setters and getters.
Reorganized the output registration in FairMCApplication::RegisterOutput().
Unified the names of the fOutFolder, used to automatically create the tree,
on the worker threads in the FairRootFileSink.
Added comment to the changelog.